backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
openai-whisper==20231117
numpy==1.26.4
ffmpeg-python==0.2.0
torch>=2.0.0
